@@ -61,31 +61,31 @@
     "ci:project:storybook:e2e": "start-server-and-test 'serve -l 6006 packages/boilerplate-styleguide/storybook-static' http://localhost:6006 'npm run workspace:styleguide test:e2e'"
   },
   "dependencies": {
-    "@commitlint/cli": "17.6.6",
-    "@commitlint/config-conventional": "17.6.6",
+    "@commitlint/cli": "17.7.2",
+    "@commitlint/config-conventional": "17.7.0",
     "@gnowth/eslint-config": "0.0.0",
     "@gnowth/jest-config": "0.0.0",
     "@gnowth/prettier-config": "0.0.0",
     "@gnowth/typescript-config": "0.0.0",
     "@tinacms/cli": "1.5.22",
-    "concurrently": "8.2.0",
+    "concurrently": "8.2.1",
     "github-markdown-css": "5.2.0",
     "husky": "8.0.3",
-    "jest-environment-jsdom": "29.6.1",
-    "lerna": "7.1.1",
-    "lint-staged": "13.2.3",
+    "jest-environment-jsdom": "29.7.0",
+    "lerna": "7.3.0",
+    "lint-staged": "14.0.1",
     "netlify-cms-proxy-server": "1.3.24",
-    "npm-check-updates": "16.10.15",
-    "serve": "14.2.0",
-    "start-server-and-test": "2.0.0",
+    "npm-check-updates": "16.14.5",
+    "serve": "14.2.1",
+    "start-server-and-test": "2.0.1",
     "surge": "0.23.1",
     "react-router-dom": "5.3.4",
     "ts-jest": "29.1.1",
-    "turbo": "1.10.7"
+    "turbo": "1.10.15"
   },
   "devDependencies": {
-    "eslint": "8.44.0",
-    "prettier": "3.0.0",
+    "eslint": "8.51.0",
+    "prettier": "3.0.3",
     "typescript": "5.2.2"
   },
   "scripts": {
