@@ -1,8 +1,8 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { ApplicationAuth } from '@gnowth/app-auth'
 import { ApplicationPages, PageNotAuthorised, PageNotFound, PageNotPermitted } from '@gnowth/app-pages'
 import { AppApplicationLazy, AppEnvironment, AppRedirect, TokenError } from '@gnowth/lib-react'
+import { createRoot } from 'react-dom/client'
 
 import { ViewFrameDefault } from './views/view-frame-default'
 import { ViewFrameRecipes } from './views/view-frame-recipes'
@@ -26,7 +26,8 @@ setup(settings)
 // mocking flag can be added here as well or per data source or should it be a separate server could still use process.ENV
 // Note: because of the limitation of how react router work, path need to be added explicitly for Switch to work. Solution could probably be available in react-router v6
 // TODO: handle page redirect and auth
-ReactDOM.render(
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+createRoot(document.getElementById('main')!).render(
   <AppEnvironment
     boundaries={{
       [TokenError.api401]: PageNotAuthorised,
@@ -48,5 +49,4 @@ ReactDOM.render(
 
     <AppRedirect application={TokenApplication.pages} />
   </AppEnvironment>,
-  document.getElementById('main'),
 )
